news/permissions.py
from rest_framework.permissions import BasePermission, SAFE_METHODS


class ReadOnlyPermission(BasePermission):
    def has_permission(self, request, view):
        return request.method in SAFE_METHODS


class IsAuthorPermission(BasePermission):
    def has_permission(self, request, view):
        if request.method in SAFE_METHODS:
            return True
        if bool(request.user and request.user.is_authenticated and
                  request.user.author):
            return True
        else:
            return False

    def has_object_permission(self, request, view, obj):
        if request.method in SAFE_METHODS:
            return True
        elif obj.author and request.user.author:
            return True
